requirements.txt
eth_account==0.13.5
python-dotenv==1.1.0
web3==7.8.0
walletrpcs